requirements.txt
click==8.1.7
psutil==5.9.6
python-dotenv==1.0.0
schedule==1.2.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.107.0